change *RendererIT to *RendererTest; and refactor out common time range
