+++ test/java/com/arpnetworking/metrics/portal/reports/impl/chrome/PdfScreenshotRendererTest.java
@@ -13,22 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.arpnetworking.metrics.portal.reports.impl.chrome.grafana;
+package com.arpnetworking.metrics.portal.reports.impl.chrome;
 
 import com.arpnetworking.metrics.portal.TestBeanFactory;
-import com.arpnetworking.metrics.portal.reports.impl.chrome.BaseChromeIT;
-import com.arpnetworking.metrics.portal.reports.impl.chrome.grafana.testing.Utils;
 import com.arpnetworking.metrics.portal.reports.impl.testing.MockRenderedReportBuilder;
 import com.typesafe.config.Config;
 import models.internal.TimeRange;
-import models.internal.impl.GrafanaReportPanelReportSource;
 import models.internal.impl.PdfReportFormat;
+import models.internal.impl.WebPageReportSource;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 
 import java.net.URI;
-import java.time.Duration;
 import java.time.Instant;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.TimeUnit;
@@ -39,13 +36,13 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static org.junit.Assert.assertTrue;
 
 /**
- * Tests class {@link PdfGrafanaScreenshotRenderer}.
+ * Tests class {@link PdfScreenshotRenderer}.
  *
- * This test is ignored on systems where it can't find Chrome -- see {@link BaseChromeIT} for instructions for manual execution.
+ * This test is ignored on systems where it can't find Chrome -- see {@link BaseChromeTest} for instructions for manual execution.
  *
  * @author Spencer Pearson (spencerpearson at dropbox dot com)
  */
-public class PdfScreenshotRendererIT extends BaseChromeIT {
+public class PdfScreenshotRendererTest extends BaseChromeTest {
 
     @Test
     public void testRendering() throws Exception {
@@ -55,23 +52,20 @@ public class PdfScreenshotRendererIT extends BaseChromeIT {
         _wireMock.givenThat(
                 get(urlEqualTo("/"))
                         .willReturn(aResponse()
-                                .withHeader("Content-Type", "text/html")
-                                .withBody(Utils.mockGrafanaReportPanelPage(Duration.ZERO))
+                                .withBody("here are some bytes")
                         )
         );
+
         final PdfReportFormat format = new PdfReportFormat.Builder().setWidthInches(8.5f).setHeightInches(11f).build();
-        final PdfGrafanaScreenshotRenderer renderer = new PdfGrafanaScreenshotRenderer(config, _renderService, _timeoutService);
-        final GrafanaReportPanelReportSource source = new GrafanaReportPanelReportSource.Builder()
-                .setWebPageReportSource(
-                        TestBeanFactory.createWebPageReportSourceBuilder()
-                                .setUri(URI.create("http://localhost:" + _wireMock.port()))
-                                .build())
+        final PdfScreenshotRenderer renderer = new PdfScreenshotRenderer(config, _renderService, _timeoutService);
+        final WebPageReportSource source = TestBeanFactory.createWebPageReportSourceBuilder()
+                .setUri(URI.create("http://localhost:" + _wireMock.port()))
                 .build();
 
         final CompletionStage<MockRenderedReportBuilder> stage = renderer.render(
                 source,
                 format,
-                new TimeRange(Instant.EPOCH, Instant.EPOCH),
+                DEFAULT_TIME_RANGE,
                 builder,
                 DEFAULT_TIMEOUT
         );
